Add button display capes.
This commit adds a bit of data to the model in the form of a 'recommendation' to the original route of flight. It adds some display capability to give the user of the app visibility into which of the planned routes actually have contrail interference, and then changes the user interface to dispaly to them the ability to modify a route of flight.