Add about page.
This adds an 'about' page using a vue component. It adds a bit of css to the original file, but it doesn't work that great. Also, the text itself might be better kept in a different source file, and loaded into the about page separately. There is also some issue with the positioning of elements on the page when the web-browser gets resized.